move SQLToken to rewrite package
