+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/token/pojo/EncryptColumnToken.java
@@ -20,7 +20,6 @@ package org.apache.shardingsphere.core.rewrite.token.pojo;
 import lombok.Getter;
 import lombok.ToString;
 import org.apache.shardingsphere.core.parse.sql.context.condition.Column;
-import org.apache.shardingsphere.core.parse.sql.token.SQLToken;
 
 /**
  * Encrypt column token.
