+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/token/pojo/OrderByToken.java
@@ -17,8 +17,6 @@
 
 package org.apache.shardingsphere.core.rewrite.token.pojo;
 
-import org.apache.shardingsphere.core.parse.sql.token.SQLToken;
-
 /**
  * Order by token.
  *
