+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/token/pojo/SelectItemsToken.java
@@ -20,7 +20,6 @@ package org.apache.shardingsphere.core.rewrite.token.pojo;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
-import org.apache.shardingsphere.core.parse.sql.token.SQLToken;
 
 import java.util.Collection;
